Ignore failures for setting the repeat rate
If the device tree hasn't been setup for repeat rate modifications, this
fails. Unfortunately this is mostly a hidden failure, but that's the
best we can do for now. Exiting with an error is really hard to debug.
Having people tell you that the repeat rate wasn't modified was easier.